Atomikos config for postgresql and mysql databases

+++ src/main/resources/db/data/db2/ddl/1_create_table_db2.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS entity2
 (
-    Id bigint NOT NULL ,
+    Id serial NOT NULL ,
     Name varchar(40) NOT NULL ,
     Balance int NOT NULL ,
     PRIMARY KEY (Id)
